refactor: home components into smaller slices
