+++ web/src/pages/Home/components/Sections/benefits/styled.tsx
@@ -1,0 +1,15 @@
+import styled from 'styled-components';
+
+export const Illustration = styled.img`
+width: 632.24px;
+height: 549px;
+
+@media (max-width: 1024px) {
+  width: 100vw;
+}
+`;
+
+export const BenefitIcon = styled.img`
+width: 78px;
+height: 78px;
+`;
