--- web/src/pages/Home/components/Sections/hero/HeroTitle.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components';
-
-const HeroTitle = styled.span`
-font-size: 90px;
-font-weight: bold;
-color: #2B2E3C;
-@media (max-width: 1024px) {
-  font-size: 40px;
- text-align: left;
-  }
-`;
-
-export default HeroTitle;
