+++ web/src/pages/Home/components/Sections/news/NewsSection.tsx
@@ -1,52 +1,11 @@
-import styled from 'styled-components';
-import PrimaryButton from '../../../../../components/buttons/PrimaryButton';
-
-import PersonIllustration from '../../../../../components/illustrations/PersonIllustration';
-import ArrowRight from '../../../../../components/icons/ArrowRight';
 import Heading2 from '../../../../../components/text/Heading2';
-
-const NewsIcon = styled(PersonIllustration)`
-  position: absolute;
-  right: 0;
-  width: 150.03px;
-  height: 138.7px;
-  top: -120px;
-  @media (max-width: 1024px) {
-    /* margin-top: 4rem; */
-    width: 97px;
-    height: 90px;
-  }
-`;
-
-const Centered = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-interface ReadMoreProps { href: string }
-
-function ReadMore({ href }: ReadMoreProps) {
-  return (
-    <div className="is-flex">
-      <a className="is-uppercase has-text-weight-bold" href={href}>
-        Read more
-      </a>
-      <ArrowRight innerFill="#5397f5" />
-    </div>
-  );
-}
+import NewsEntry from '.';
+import { Centered, NewsIcon, ReadMoreNews } from './styled';
+import ReadMore from './ReadMore';
 
 function FeaturedEntry() {
   return (
     <div className="card">
-      {/* <div className="card-image">
-        <figure className="image is-4by3">
-          <img
-            src="https://bulma.io/images/placeholders/1280x960.png"
-            alt="Placeholder image"
-          />
-        </figure>
-      </div> */}
       <div style={{ background: '#F8F8F8' }} className="card-content">
         <div className="media">
           <div className="media-content">
@@ -65,25 +24,6 @@ function FeaturedEntry() {
     </div>
   );
 }
-
-interface NewsEntryProps {
-  title: string;
-  description: string;
-  link: string;
-}
-function NewsEntry({ title, description, link }: NewsEntryProps) {
-  return (
-    <div className="card">
-      <div className="card-content">
-        <h3 className="title is-4">{title}</h3>
-        <p className="content">{description}</p>
-        <ReadMore href={link} />
-      </div>
-    </div>
-  );
-}
-
-const ReadMoreNews = styled(PrimaryButton)``;
 
 function NewsSection() {
   return (
